fix: 修复体验问题
